#29 [MODIFY] question 모두 Integer 타입으로 변경

+++ src/main/java/healeat/server/repository/MemberHealQuestionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberHealQuestionRepository extends JpaRepository<MemberHealQuestion, Long> {
+public interface MemberHealQuestionRepository extends JpaRepository<MemberHealQuestion, Integer> {
 }
